Improve output with depth-awareness and single-property object collapsing
Objects can now return a reduced (or otherwise different) representation of themselves when view at different depths. e.g. if an object is being dumped as a grandchild of the main target object, we don't necessarily need all the data. But if it itself is the target, it should return everything.

Also, there's a centralised method to add a value to the output object. Objects with a single property will now have that property collapsed onto the parent property, with keys combined.

e.g. this input:
{"CMS": {"version": "1.12.1.15"} }
will be output as if it is
{"CMS version": "1.12.1.15"}